+++ packages/db/migrations/0005_moaning_virginia_dare.sql
@@ -1,0 +1,1 @@
+ALTER TYPE "public"."integration_kind" ADD VALUE 'http';